// src/app/models/profileInfo.ts
export class ProfileInfo {
    profileID: number;
    userID: number;
    pFullName: string;
    pTitle: string;
    pDescription: string;
    pSittingCapacity: string;
    pCompaign: string;
}
export class PortfolioInfo {
    portFolioID:number;
    userID: number;
    profileID: number;
    portFolioPath: string;
}